code/client/requirements.txt
Faker==18.11.2
python-dateutil==2.8.2
pyzmq==25.1.0
six==1.16.0
